# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quablog API - Q&A and Blogging Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (self or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/{id}/block:
    post:
      operationId: blockUser
      tags: [users]
      summary: Block/unblock a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserRequest"
      responses:
        "200":
          description: User blocked/unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── QUESTIONS ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # Auth
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            questions:
              type: integer
            answers:
              type: integer
            blogs:
              type: integer
          nullable: true
      required: [id, username, email, displayName, role, isBlocked, createdAt]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        _count:
          type: object
          properties:
            questions:
              type: integer
            answers:
              type: integer
            blogs:
              type: integer
      required: [id, username, displayName, role, createdAt, _count]
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [users, total, page, limit, totalPages]
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    BlockUserRequest:
      type: object
      properties:
        blocked:
          type: boolean
      required: [blocked]

    # Questions
...